Don't flip image by default
In Teams the background image as it appears
to you is mirrored; however to the other
members of the meeting the image is
correctly oriented. The mirroring that
was implemented here made the image look correct
to the user, but incorrectly flipped for others.
This change makes mirroring optional,
defaulting to off.